# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flexi Route Logistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List shipments (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, in_transit, out_for_delivery, delivered, cancelled]
        - name: search
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: driverId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentListResponse"

    post:
      operationId: createShipment
      tags: [shipments]
      summary: Book a shipment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentInput"
      responses:
        "201":
          description: Shipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get shipment by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shipment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateShipment
      tags: [shipments]
      summary: Update shipment (admin/driver)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentUpdate"
      responses:
        "200":
          description: Updated shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"

    delete:
      operationId: deleteShipment
      tags: [shipments]
      summary: Delete shipment (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /shipments/{id}/assign:
    post:
      operationId: assignDriver
      tags: [shipments]
      summary: Assign a driver to shipment (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverAssignment"
      responses:
        "200":
          description: Driver assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /shipments/{id}/events:
    get:
      operationId: getShipmentEvents
      tags: [shipments]
      summary: Get tracking events for shipment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tracking events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackingEvent"

    post:
      operationId: addTrackingEvent
      tags: [shipments]
      summary: Add tracking event (admin/driver)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingEventInput"
      responses:
        "201":
          description: Event added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingEvent"
  /shipments/{id}/proof:
    post:
      operationId: uploadDeliveryProof
      tags: [shipments]
      summary: Upload proof of delivery (driver)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ProofUpload"
      responses:
        "200":
          description: Proof uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"

  # --- PUBLIC TRACKING ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # --- AUTH SCHEMAS ---
...
    Shipment:
      type: object
      required: [id, trackingNumber, status, serviceType, originAddress, destinationAddress, createdAt]
      properties:
        id:
          type: integer
        trackingNumber:
          type: string
        customerId:
          type: integer
        driverId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, processing, in_transit, out_for_delivery, delivered, cancelled]
        serviceType:
          type: string
          enum: [standard, express, overnight, freight]
        originAddress:
          type: string
        originCity:
          type: string
        originState:
          type: string
        originZip:
          type: string
        destinationAddress:
          type: string
        destinationCity:
          type: string
        destinationState:
          type: string
        destinationZip:
          type: string
        weightKg:
          type: number
        dimensions:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
          format: date
        deliveryProofUrl:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ShipmentDetail:
      allOf:
        - $ref: "#/components/schemas/Shipment"
        - type: object
          properties:
            events:
              type: array
              items:
                $ref: "#/components/schemas/TrackingEvent"
            payment:
              $ref: "#/components/schemas/Payment"
    ShipmentInput:
      type: object
      required: [serviceType, originAddress, originCity, originState, originZip, destinationAddress, destinationCity, destinationState, destinationZip, weightKg]
      properties:
        serviceType:
          type: string
          enum: [standard, express, overnight, freight]
        originAddress:
          type: string
        originCity:
          type: string
        originState:
          type: string
        originZip:
          type: string
        destinationAddress:
          type: string
        destinationCity:
          type: string
        destinationState:
          type: string
        destinationZip:
          type: string
        weightKg:
          type: number
        dimensions:
          type: string
        description:
          type: string
        senderName:
          type: string
        senderPhone:
          type: string
        recipientName:
          type: string
        recipientPhone:
          type: string
    ShipmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, in_transit, out_for_delivery, delivered, cancelled]
        driverId:
          type: ["integer", "null"]
        estimatedDelivery:
          type: string
          format: date
        totalAmount:
          type: number
    ShipmentListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Shipment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DriverAssignment:
      type: object
      required: [driverId]
      properties:
        driverId:
          type: integer
    ProofUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        notes:
          type: string

    # --- TRACKING SCHEMAS ---
    TrackingEvent:
      type: object
      required: [id, shipmentId, status, description, createdAt]
      properties:
        id:
          type: integer
        shipmentId:
          type: integer
        status:
          type: string
        location:
          type: ["string", "null"]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    TrackingEventInput:
      type: object
      required: [status, description]
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, in_transit, out_for_delivery, delivered, cancelled]
        location:
          type: string
        description:
          type: string
...
    Payment:
      type: object
      required: [id, shipmentId, amount, currency, walletAddress, status, createdAt]
      properties:
        id:
          type: integer
        shipmentId:
          type: integer
        amount:
          type: number
        currency:
          type: string
          enum: [BTC, ETH, USDT_TRC20, USDT_ERC20, USDC, LTC]
        walletAddress:
          type: string
        txid:
          type: ["string", "null"]
        paymentProofUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [awaiting_payment, under_review, confirmed, rejected]
        adminNotes:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...